+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "al-codeactions" extension will be documented in this
 
 - Minor changes
   - Create procedure in argument list even if the procedure does not exist yet
+  - Create Procedure in Validate Statement
   - Create overload of procedure: fix if the new parameter is a text variable
 
 ## 1.0.2